// src/app/models/permissions.model.ts
export enum Permission {
    GROUP_VIEW = 'group:view',
    GROUP_EDIT = 'group:edit',
    GROUP_ADD = 'group:add',
    GROUP_DELETE = 'group:delete',

    TICKET_VIEW = 'ticket:view',
    TICKETS_VIEW = 'tickets:view',
    TICKET_EDIT = 'ticket:edit',
    TICKET_ADD = 'ticket:add',
    TICKET_DELETE = 'ticket:delete',
    TICKET_EDIT_STATE = 'ticket:edit_state',

    USER_VIEW = 'user:view',
    USERS_VIEW = 'users:view',
    USER_EDIT = 'user:edit',
    USER_ADD = 'user:add',
    USER_DELETE = 'user:delete',

    PERMISSIONS_MANAGE = 'permissions:manage'
}